wip
switched to closest-bot detection. Kind of happy with the mechanics of the mod, this version needs to be cleaned a lot but is in release shape. Want to add switched for luma and breadth, as well as performance mod. Mod functions well with just breadth with closest-bot at 1 tick per second, very good performance and result which can all function with ThatsLit in tandem.